chore: reorganize files
Put all sources into "src" folder
