--- __tests__/outputs/rnuc.xcconfig
@@ -1,8 +1,0 @@
-// DO NOT COMMIT OR EDIT THIS FILE
-MY_VARIABLE=hello
-MY_URL=http://hello.world?howareyoudoing=ok
-MY_STRING=string_value
-MY_NUMBER=42
-MY_BOOLEAN=true
-MY_BOOLEAN_FALSE=false
-PER_PLATFORM=hello
