+++ src/templates/index.d.ts.handlebars
