+++ src/templates/rnuc.xcconfig.handlebars
