+++ src/write-env.js
@@ -1,8 +1,7 @@
 const fs = require("fs");
 
-function write_env(env) {
+module.exports = function (env) {
   for (const path in env) {
     fs.writeFileSync(path, env[path]);
   }
-}
-module.exports = write_env;
+};
